Store data instead of cell instances
Only store data representation of cell instead of cell instances. The
cell class should have a get data method and a way to instantiate from
existing data. The grid will now need to be responsible for carving
walls on cells since the cell instances used in the main carve maze
function will no longer be references to instances in the grid's cell
array. The grid will need to get the cell data, mutate it, and then add
it back to the cells list.